# [fbcode][static runtime] out-variant for quantized::linear_dynamic_fp16